# doc/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "c2ae-api",
    "version": "version not set"
  },
  "host": "127.0.0.1:8886",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/rules": {
      "get": {
        "summary": "Retrieve list of existing rules",
        "operationId": "ListRules",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbRulesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "tags": [
          "C2AutomationEngine"
        ]
      },
      "post": {
        "summary": "Create a new rule",
        "operationId": "AddRule",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbRuleResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbAddRuleRequest"
            }
          }
        ],
        "tags": [
          "C2AutomationEngine"
        ]
      },
      "put": {
        "summary": "Update an existing rule",
        "operationId": "UpdateRule",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbRuleResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbUpdateRuleRequest"
            }
          }
        ],
        "tags": [
          "C2AutomationEngine"
        ]
      }
    },
    "/rules/{ruleId}": {
      "get": {
        "summary": "Retrieve a single rule, by its ID",
        "operationId": "GetRule",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbRuleResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "ruleId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "C2AutomationEngine"
        ]
      },
      "delete": {
        "summary": "Remove a rule",
        "operationId": "DeleteRule",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbDeleteRuleResponse"
            }
          },
          "default": {
            "description": "An unexpected error response",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "ruleId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "C2AutomationEngine"
        ]
      }
    }
  },
  "definitions": {
    "pbActionType": {
      "type": "string",
      "enum": [
        "UNDEFINED_ACTION",
        "KEY_ROTATION"
      ],
      "default": "UNDEFINED_ACTION",
      "title": "List of supported ActionType"
    },
    "pbAddRuleRequest": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string"
        },
        "action": {
          "$ref": "#/definitions/pbActionType"
        },
        "triggers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/pbTrigger"
          }
        },
        "targets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/pbTarget"
          }
        }
      }
    },
    "pbDeleteRuleResponse": {
      "type": "object",
      "properties": {
        "ruleId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "pbRule": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "description": {
          "type": "string"
        },
        "action": {
          "$ref": "#/definitions/pbActionType"
        },
        "lastExecuted": {
          "type": "string",
          "format": "date-time"
        },
        "triggers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/pbTrigger"
          }
        },
        "targets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/pbTarget"
          }
        }
      }
    },
    "pbRuleResponse": {
      "type": "object",
      "properties": {
        "rule": {
          "$ref": "#/definitions/pbRule"
        }
      }
    },
    "pbRulesResponse": {
      "type": "object",
      "properties": {
        "rules": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/pbRule"
          }
        }
      }
    },
    "pbTarget": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "$ref": "#/definitions/pbTargetType"
        },
        "expr": {
          "type": "string"
        }
      }
    },
    "pbTargetType": {
      "type": "string",
      "enum": [
        "ANY",
        "TOPIC",
        "CLIENT"
      ],
      "default": "ANY",
      "title": "List of supported TargetType"
    },
    "pbTrigger": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "$ref": "#/definitions/pbTriggerType"
        },
        "settings": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "pbTriggerType": {
      "type": "string",
      "enum": [
        "UNDEFINED_TRIGGER",
        "TIME_INTERVAL",
        "EVENT"
      ],
      "default": "UNDEFINED_TRIGGER",
      "title": "List of supported TriggerType"
    },
    "pbUpdateRuleRequest": {
      "type": "object",
      "properties": {
        "ruleId": {
          "type": "integer",
          "format": "int32"
        },
        "description": {
          "type": "string"
        },
        "action": {
          "$ref": "#/definitions/pbActionType"
        },
        "triggers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/pbTrigger"
          }
        },
        "targets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/pbTarget"
          }
        }
      },
      "description": "UpdateRuleRequest will fetch the rule identified by ruleId,\nand override its description, action, triggers and targets values\nwith those provided."
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string",
          "description": "A URL/resource name that uniquely identifies the type of the serialized\nprotocol buffer message. This string must contain at least\none \"/\" character. The last segment of the URL's path must represent\nthe fully qualified name of the type (as in\n`path/google.protobuf.Duration`). The name should be in a canonical form\n(e.g., leading \".\" is not accepted).\n\nIn practice, teams usually precompile into the binary all types that they\nexpect it to use in the context of Any. However, for URLs which use the\nscheme `http`, `https`, or no scheme, one can optionally set up a type\nserver that maps type URLs to message definitions as follows:\n\n* If no scheme is provided, `https` is assumed.\n* An HTTP GET on the URL must yield a [google.protobuf.Type][]\n  value in binary format, or produce an error.\n* Applications are allowed to cache lookup results based on the\n  URL, or have them precompiled into a binary to avoid any\n  lookup. Therefore, binary compatibility needs to be preserved\n  on changes to types. (Use versioned type names to manage\n  breaking changes.)\n\nNote: this functionality is not currently available in the official\nprotobuf release, and it is not used for type URLs beginning with\ntype.googleapis.com.\n\nSchemes other than `http`, `https` (or the empty scheme) might be\nused with implementation specific semantics."
        },
        "value": {
          "type": "string",
          "format": "byte",
          "description": "Must be a valid serialized protocol buffer of the above specified type."
        }
      },
      "description": "`Any` contains an arbitrary serialized protocol buffer message along with a\nURL that describes the type of the serialized message.\n\nProtobuf library provides support to pack/unpack Any values in the form\nof utility functions or additional generated methods of the Any type.\n\nExample 1: Pack and unpack a message in C++.\n\n    Foo foo = ...;\n    Any any;\n    any.PackFrom(foo);\n    ...\n    if (any.UnpackTo(\u0026foo)) {\n      ...\n    }\n\nExample 2: Pack and unpack a message in Java.\n\n    Foo foo = ...;\n    Any any = Any.pack(foo);\n    ...\n    if (any.is(Foo.class)) {\n      foo = any.unpack(Foo.class);\n    }\n\n Example 3: Pack and unpack a message in Python.\n\n    foo = Foo(...)\n    any = Any()\n    any.Pack(foo)\n    ...\n    if any.Is(Foo.DESCRIPTOR):\n      any.Unpack(foo)\n      ...\n\n Example 4: Pack and unpack a message in Go\n\n     foo := \u0026pb.Foo{...}\n     any, err := ptypes.MarshalAny(foo)\n     ...\n     foo := \u0026pb.Foo{}\n     if err := ptypes.UnmarshalAny(any, foo); err != nil {\n       ...\n     }\n\nThe pack methods provided by protobuf library will by default use\n'type.googleapis.com/full.type.name' as the type URL and the unpack\nmethods only use the fully qualified type name after the last '/'\nin the type URL, for example \"foo.bar.com/x/y.z\" will yield type\nname \"y.z\".\n\n\nJSON\n====\nThe JSON representation of an `Any` value uses the regular\nrepresentation of the deserialized, embedded message, with an\nadditional field `@type` which contains the type URL. Example:\n\n    package google.profile;\n    message Person {\n      string first_name = 1;\n      string last_name = 2;\n    }\n\n    {\n      \"@type\": \"type.googleapis.com/google.profile.Person\",\n      \"firstName\": \u003cstring\u003e,\n      \"lastName\": \u003cstring\u003e\n    }\n\nIf the embedded message type is well-known and has a custom JSON\nrepresentation, that representation will be embedded adding a field\n`value` which holds the custom JSON in addition to the `@type`\nfield. Example (for message [google.protobuf.Duration][]):\n\n    {\n      \"@type\": \"type.googleapis.com/google.protobuf.Duration\",\n      \"value\": \"1.212s\"\n    }"
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}